+++ Sensors.c
@@ -2,48 +2,48 @@
 //Team 57
 //  Jillian Hestle, Emily Schott, Nicholas Vilbrandt, Tyler Stagge
 
-//GENERAL SENSOR INTERPRETATION MODULE [Sensors.c]
+//GENERAL SENSOR shortERPRETATION MODULE [Sensors.c]
 //
-//This file contains the functions required to operate and interpret most
+//This file contains the functions required to operate and shorterpret most
 //sensors utilized by the ALV; see Driver.c for pragmas and global variables
 
 /*----------------------------FUNCTION  PROTOTYPES----------------------------*/
-int calibrateHallEffect();
-int getHallEffect(int base);
-bool nearBeacon(int base);
-bool atBeacon(int base);
+short calibrateHallEffect();
+short getHallEffect(short base);
+bool nearBeacon(short base);
+bool atBeacon(short base);
 
-int calibrateGyro();
-int getGyro(int base);
+short calibrateGyro();
+short getGyro(short base);
 /*----------------------------FUNCTION DEFINITIONS----------------------------*/
-int calibrateHallEffect()
+short calibrateHallEffect()
 {
 	return SensorValue(S3);
 }
 
-int getHallEffect(int base)
+short getHallEffect(short base)
 {
 	return -(SensorValue(S3) - base);
 }
 
-bool nearBeacon(int base)
+bool nearBeacon(short base)
 {
-	int thresh1 = HALL_EFFECT_THRESH_1;
+	short thresh1 = HALL_EFFECT_THRESH_1;
 	return (getHallEffect(base) >= thresh1);
 }
 
-bool atBeacon(int base)
+bool atBeacon(short base)
 {
-	int thresh3 = HALL_EFFECT_THRESH_3;
+	short thresh3 = HALL_EFFECT_THRESH_3;
 	return (getHallEffect(base) >= thresh3);
 }
 
-int calibrateGyro()
+short calibrateGyro()
 {
 	return SensorValue(S1);
 }
 
-int getGyro(int base)
+short getGyro(short base)
 {
 	//Note: for raw gyro readings, positive is right, negative is left
 	//      Thus they are being flipped to be consistent with the standard that positive is left
